requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23

# Data processing (use latest versions for Python 3.13 compatibility)
numpy>=1.26.0
pandas>=2.1.0

# API and web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0

# Logging and monitoring
colorlog>=6.8.0
